blog/app/requirements.txt
fastapi

uvicorn

sqlalchemy

passlib

bcrypt

starlette

python-jose

python-multipart
